refactor: Stage 1 of the crate consolidation - remove obsolete debris
Removes dead files and dependencies without touching runtime architecture.
Package count 34 -> 33.

- Drop the cljrs-runtime dependency from the CLI, and from its no-gc feature
  list. Nothing in the workspace depends on the stub now.
- Delete crates/cljrs-stdlib/build.rs. It built no IR; it only emitted a
  rerun-if-changed directive for itself.
- Delete crates/cljrs-compiler/src/ir.rs and rewrite the 27 crate::ir:: paths
  inside cljrs-compiler to cljrs_ir::. The module only re-exported cljrs-ir and
  drew no compiler boundary; no package outside cljrs-compiler referenced it.
- Fold cljrs-ir-prebuild into cljrs::commands::ir and delete the package. The
  standalone binary duplicated `cljrs ir build`. The commands stay: they are
  useful lowerer diagnostics, and cljrs_eval::load_prebuilt_ir remains the
  replay path for embedders with no background lowering worker. Baseline §5
  found no cold-start benefit to preserve - nothing loads a bundle today.

  This creates crates/cljrs/src/commands/, the module tree Stage 5 splits
  main.rs into, and moves IrCommands, its dispatch, and `ir viz` there.

- Correct stale documentation. The root README described a build-time IR
  bootstrap through Clojure compiler namespaces that does not exist; the
  cljrs-runtime README described a clojure.core implementation that actually
  lives in cljrs-builtins and cljrs-interp. Also refreshes the cljrs-stdlib
  file and API lists, the cljrs-compiler and cljrs README dependency tables,
  the cljrs-eval load_prebuilt_ir note, and the workspace test count.

Two of the plan's Stage 1 items needed no change: cljrs-stdlib/src/core_async.rs
does not exist, and cljrs-stdlib no longer declares tokio or lazy_static.

Validation, all matching the Stage 0 baseline: fmt, clippy -D warnings, 1077
workspace tests (144 suites vs 147 - the three cljrs-ir-prebuild targets are
gone, no test lost), the 11,005-assertion AOT Clojure suite, and the graph,
life, and core_async samples. CLI help output is byte-identical to the baseline
binary except for the `ir build` text that wrongly claimed bundles load at
startup. `cljrs ir build --ns clojure.core` lowers the same 151 functions with
the same 2 unsupported. no-gc is unchanged from baseline: still broken in the
same four packages, pre-existing and out of scope here.
