+++ Cargo.toml
@@ -17,7 +17,6 @@ members = [
     "crates/cljrs-env",
     "crates/cljrs-builtins",
     "crates/cljrs-interp",
-    "crates/cljrs-ir-prebuild",
     "crates/cljrs-ir-viz",
     "crates/cljrs-deps",
     "crates/cljrs-vcs",
@@ -57,7 +56,6 @@ cljrs-logging      = { path = "crates/cljrs-logging",      version = "0.1.0" }
 cljrs-env          = { path = "crates/cljrs-env",          version = "0.1.0" }
 cljrs-builtins     = { path = "crates/cljrs-builtins",     version = "0.1.0" }
 cljrs-interp       = { path = "crates/cljrs-interp",       version = "0.1.0" }
-cljrs-ir-prebuild  = { path = "crates/cljrs-ir-prebuild",  version = "0.1.0" }
 cljrs-ir-viz       = { path = "crates/cljrs-ir-viz",       version = "0.1.0" }
 cljrs-deps         = { path = "crates/cljrs-deps",         version = "0.1.0" }
 cljrs-vcs          = { path = "crates/cljrs-vcs",          version = "0.1.0" }
